Implemented `Engine::with_*_from`.
This allows slightly easier initialisation from a source sequence.  Unlike
the original C++ implementation, this uses `Iterator` rather than a special
RNG seed type.

The *downside* of this is that an RNG could be initialised with fairly
rubbish data.  Also, it will require *slightly* more work to initialise
with a non-PCG RNG.

The *upside* is that `Iterator` is about as general as possible.